menu: added menu selector component
