+++ src/Components/Body/testimonial.js
@@ -1,97 +1,64 @@
-import React from 'react'
-import Carousel from "react-multi-carousel";
+import React, { useState } from "react";
+import { Carousel } from "react-bootstrap";
 
-function Testimonial() {
+const useStyle = () => ({
+  container: {
+    padding: "30px 60px",
+    background: "#272727",
+  },
+  header: {
+    fontFamily: "Clicker Script",
+    fontSize: "10vw",
+    color: "#B29051",
+  },
+  carouselContent: {
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    color: "white",
+    minHeight: "50vh",
+    width: "100%",
+  },
+  carouselMessage: {
+    width: "70%",
+  },
+});
 
-    const testimonialsArr=[
-        {
-            name:'Emanuele S',
-            date:'October 1 2020',
-            text:'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque nisl eros, \n' +
-                'pulvinar facilisis justo mollis'
-        },
-        {
-            name:'Emanuele S',
-            date:'October 1 2020',
-            text:'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque nisl eros, \n' +
-                'pulvinar facilisis justo mollis'
-        },
-        {
-            name:'Emanuele S',
-            date:'October 1 2020',
-            text:'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque nisl eros, \n' +
-                'pulvinar facilisis justo mollis'
-        }
-    ]
-    return(
-        <div className='w-full '>
-            <Carousel
-                additionalTransfrom={0}
-                arrows
-                autoPlaySpeed={3000}
-                centerMode={false}
-                className=""
-                containerClass="carousel-container"
-                dotListClass=""
-                draggable
-                focusOnSelect={false}
-                infinite
-                itemClass="carousel-item-padding-30-px"
-                keyBoardControl
-                minimumTouchDrag={80}
-                renderButtonGroupOutside={false}
-                renderDotsOutside={false}
-                responsive={{
-                    desktop: {
-                        breakpoint: {
-                            max: 3000,
-                            min: 1024
-                        },
-                        items: 2,
-                        partialVisibilityGutter: 40
-                    },
-                    mobile: {
-                        breakpoint: {
-                            max: 464,
-                            min: 0
-                        },
-                        items: 1,
-                        partialVisibilityGutter: 30
-                    },
-                    tablet: {
-                        breakpoint: {
-                            max: 1024,
-                            min: 464
-                        },
-                        items: 1,
-                        partialVisibilityGutter: 30
-                    }
-                }}
-                showDots={false}
-                sliderClass=""
-                slidesToSlide={1}
-                swipeable
-            >
+export default function Testimonial() {
+  const styles = useStyle();
+  const [testimonials, setTestimonials] = useState([
+    {
+      img: "assets/testimonial-client.png",
+      content:
+        "Birre ottime e panini buoni nel pane e nella farcitura, carne scelta. Locale informale e caldo. Ragazzi preparati e gentili. Patatine buone servite con più salse. Buoni i gelati e i sorbettie buono il soufflé al cioccolato. Da provare indubbiamente",
+      author: "Giulia",
+    },
+    {
+      img: "assets/testimonial-client.png",
+      content:
+        "Birre ottime e panini buoni nel pane e nella farcitura, carne scelta. Locale informale e caldo. Ragazzi preparati e gentili. Patatine buone servite con più salse. Buoni i gelati e i sorbettie buono il soufflé al cioccolato. Da provare indubbiamente",
+      author: "Giulia",
+    },
+  ]);
 
-                {
-                    testimonialsArr.map((item, index) => (
-                        <div key={index} className="max-w-md py-4 border-1 border-gray-700  px-8 bg-white ml-2 rounded-lg my-20">
-                            <div className="flex justify-center md:justify-end -mt-16">
-                                <img className="w-20 h-20 object-cover rounded-full "
-                                     src="https://images.unsplash.com/photo-1499714608240-22fc6ad53fb2?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=334&q=80"/>
-                            </div>
-                            <div className='border-1 border-gray-700'>
-                                <h2 className="text-gray-800 text-lg font-semibold">{item.name}</h2>
-                                <p className="mt-2 text-xs text-gray-600">{item.date}</p>
-                                <p className="mt-2 text-xs text-gray-600">{item.text}</p>
-                            </div>
-
-                        </div>
-                    ))
-                }
-
-            </Carousel>
-        </div>)
+  return (
+    <div style={styles.container}>
+      <h1 style={styles.header}>Our client says</h1>
+      <Carousel interval={1000}>
+        {testimonials.map((testimonial) => {
+          return (
+            <Carousel.Item interval={5000}>
+              <div style={styles.carouselContent}>
+                <img src={testimonial.img} className="mb-3" />
+                <p style={styles.carouselMessage}>‘’{testimonial.content}’’</p>
+              </div>
+              <Carousel.Caption>
+                <p>-{testimonial.author}-</p>
+              </Carousel.Caption>
+            </Carousel.Item>
+          );
+        })}
+      </Carousel>
+    </div>
+  );
 }
-
-export default Testimonial
