+++ src/Components/Main/reserveTableMenuList.css
@@ -11,17 +11,3 @@
   > div::-webkit-scrollbar-thumb {
   background-color: yellow;
 }
-
-#root
-  > div.App
-  > div
-  > section:nth-child(2)
-  > div
-  > div
-  > div.col-12.col-lg-6.d-flex.flex-column.px-0.justify-content-stretch
-  > div:nth-child(2)
-  > div
-  > div
-  > div::-webkit-scrollbar-track {
-  background-color: yellow;
-}
