+++ src/Components/Menu/MenuContent.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function MenuContent() {
+  return (
+    <div>
+      <h1>Menu Content</h1>
+    </div>
+  );
+}
