+++ src/Components/Menu/menuSelector.css
@@ -1,0 +1,32 @@
+#root > div.App > div > section:nth-child(2) > div > div > div > div > div {
+  margin: 10px 20px;
+}
+#root
+  > div.App
+  > div
+  > section:nth-child(2)
+  > div
+  > div
+  > div
+  > div
+  > div
+  > button {
+  padding: 1.1em;
+  color: white;
+}
+
+#root
+  > div.App
+  > div
+  > section:nth-child(2)
+  > div
+  > div
+  > div
+  > div
+  > div
+  > button.active {
+  transform: scale(1.2);
+  margin: 5px;
+  background-color: #272727;
+  box-shadow: -2px 4px 16px 0px rgba(0, 0, 0, 0.75);
+}
